Move world interaction stuff into own dir
