+++ src/file_system_interaction/game_serialization.rs
@@ -1,8 +1,8 @@
-use crate::condition::ActiveConditions;
-use crate::dialog::{CurrentDialog, DialogEvent};
 use crate::file_system_interaction::level_serialization::{CurrentLevel, WorldLoadRequest};
 use crate::movement_gameplay::player::Player;
 use crate::spawning::{DelayedSpawnEvent, GameObject, SpawnEvent};
+use crate::world_interaction::condition::ActiveConditions;
+use crate::world_interaction::dialog::{CurrentDialog, DialogEvent};
 use crate::GameState;
 use bevy::prelude::*;
 use chrono::prelude::Local;
