+++ src/world_interaction/condition.rs
